Move UserArgs to a types module
